backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt==4.3.0",
    "fastapi[standard]>=0.128.0",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "passlib==1.7.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.45",
    "valkey>=6.1.1",
]
